Change landmark ids to 32 bit unsigned

+++ beam_containers/include/beam_containers/LandmarkMeasurement.h
@@ -30,7 +30,7 @@ template <typename SensorIdType>
 struct LandmarkMeasurement {
   TimePoint time_point;
   SensorIdType sensor_id;
-  uint64_t landmark_id;
+  uint32_t landmark_id;
   size_t image;
   Eigen::Vector2d value;
   cv::Mat descriptor;
